int instead of float input

+++ zad1/app.py
@@ -2,8 +2,8 @@ import Calculator
 
 while True:
     try:
-        a = float(input("Input first number: "))
-        b = float(input("Input second number: "))
+        a = int(input("Input first number: "))
+        b = int(input("Input second number: "))
         print(f"NWD { Calculator.getGreatestCommonDivisor(a, b) } ")
         print(f"NWW { Calculator.getLeastCommonMultiple(a, b) } ")
